server/requirements.txt
flask>=3.0
python-dotenv>=1.0
google-generativeai>=0.8
pillow>=10.0
